+++ CHANGELOG.md
@@ -1,3 +1,5 @@
 - Mod namespace tooltips can now be disabled.
 - Mod namespace tooltips now respect `getCreatorNamespace`.
 - Hide EIV keybind is now unset by default.
+- Added buttons to change the current page.
+- Added a "Search ItemView" hint to the search bar.